HierarchyRepeater: Nest HierarchyRepeaterLevel inside HR
