+++ src/Framework/Framework/Controls/HierarchyRepeater.cs
@@ -11,6 +11,7 @@ using DotVVM.Framework.Compilation.Javascript;
 using DotVVM.Framework.Controls;
 using DotVVM.Framework.Hosting;
 using DotVVM.Framework.ResourceManagement;
+using DotVVM.Framework.Utils;
 
 namespace DotVVM.Framework.Controls
 {
@@ -384,6 +385,47 @@ namespace DotVVM.Framework.Controls
             var tempContainer = new DataItemContainer { Parent = this, DataContext = item };
             tempContainer.SetDataContextTypeFromDataSource(GetDataSourceBinding());
             return ItemChildrenBinding!.Evaluate(tempContainer) ?? Enumerable.Empty<object>();
+        }
+
+        /// <summary>
+        /// A container for a level of the <see cref="HierarchyRepeater"/>.
+        /// </summary>
+        private class HierarchyRepeaterLevel : DotvvmControl
+        {
+            public string? ItemTemplateId { get; set; }
+
+            public string? ForeachExpression { get; set; }
+
+            public bool IsRoot { get; set; } = false;
+
+            protected override void RenderControl(IHtmlWriter writer, IDotvvmRequestContext context)
+            {
+                if (ForeachExpression is null)
+                {
+                    base.RenderControl(writer, context);
+                    return;
+                }
+
+                if (!RenderOnServer)
+                {
+                    var koGroup = new KnockoutBindingGroup {
+                        { "foreach", ForeachExpression }
+                    };
+                    koGroup.AddValue("name", ItemTemplateId.NotNull());
+                    koGroup.AddValue("hierarchyRole", IsRoot ? "Root" : "Child");
+                    writer.WriteKnockoutDataBindComment("template", koGroup.ToString());
+                }
+                else
+                {
+                    writer.WriteKnockoutDataBindComment("dotvvm-SSR-foreach", new KnockoutBindingGroup {
+                        { "data", ForeachExpression}
+                    }.ToString());
+                }
+
+                base.RenderControl(writer, context);
+
+                writer.WriteKnockoutDataBindEndComment();
+            }
         }
     }
 }
